pointing_poker/aws/resources/requirements.txt
aws-cdk.core==1.36.1
aws-cdk.aws_dynamodb==1.36.1
aws-cdk.aws_appsync==1.36.1
